fix: remove lookup for states
Field "province" in Creditsafe response correspond to then state
information in Odoo.

Method get_state tried to find the state in res_country_state by name. This
method is error prone. For instance, Spanish state Malaga is returned
as 'MALAGA' in province field and stored as 'Málaga' in Odoo.

The better is to never populate the fields.

BTW, a global review of Creditsafe_data_mixin should be done, since the
same issue can happen for other list based fields.